Avoid array out of bounds panic parsing termInfo
If an offset in the string section runs off the end of the string table then
the previous code would panic with an array out of bounds issue.

Restructure the code to check for an initial offset which is too large and use
bytes.IndexByte to hunt for the nul and handle the case where one is not found.